test: move mocks to common file for integration tests
